gen: add +commit-event
Produces a unix commit event from the provided path. This is useful in situations
where a pier is running a buggy version of a hoon file that we cannot recover from.

By making a jammed event of the commit, we can inject into the affected pier
(using the "-I event.jam" flag) a restored version of the file that fixes the issue.