ames: wip support for comet attestation
Instead of sending a special packet (i.e. open-packet in old %ames) we just
read the attestation proof from the comet's namespace. There needs to be
a special case for %alien comets that we pretend to know, so we can add the
path into the pit to track it—this is a bit of a hack and might be fragile because
of the way we need to check that we have bunted the state of the chum comet

(note: untested)