eyre: explicit 401 instead of implicit downgrade
With the addition of guest identities (#6561) eyre started granting
guest identities to unauthenticated requesters. However, requesters that
tried authenticating with _invalid_ credentials (for non-existent or
expired sessions) would be granted a guest identity as well, implicitly
"downgrading" their request in the process, to appear as if originating
from the guest identity.

This behavior is neither correct nor legible. Especially API-level
interactions will want to avoid accidentally opening a channel with an
unexpected identity.

We improve Eyre's behavior by detecting the "invalid session" case and
returning an explicit 401 (unauthorized) response. To make sure that the
browser case doesn't stay stuck in a broken session, we include a
`set-cookie` header that expires the invalid cookie instantly.

In practice, this change _should_ only be user-facing in the "log out of
all sessions" case, where users may see a 401 page in browsers on their
other devices after performing that action. A simple refresh of the page
gets them back into the usual, endpoint-defined behavior.