mesa: don't migrate if there are any live messages
if we are in the middle of sending a message,
assemble fragments will crash since for any acked
fragment it has been removed from the queue.

this will cancel the migration for both sender
(when acked) and receiver of the ahoy plea.

- for the sender, the ahoy plea when resent,
will always be acked by the receiver if they
succesfully migrated us, and hearing ack
will trigger the migration.

- for the receiver, if they are sending or receiving
anything, they will nack the ahoy plea and cancel
the migration.

(note: better not to crash on the receiver of the ahoy plea
and use retry-pump metrics for the retries?)