eyre: ensure initial channel response has a body
Some clients will not consider the connection as being established until
they've received some body bytes from it. If the client has logic that
awaits the connection, this will result in a mandatory 20 second wait
(until the first heartbeat fires).

Here, if we detect that we're not sending any event backlog, we simply
send a heartbeat (empty sse comment) instead.