mesa: fix %ahoy migration
naxplanation flows are not migrated (backward and %ack forward).
the bacward side is converte into a +peek; the forward dropped.

when regressing, we get the last naxplanation and create the message
pump anew, ignoring any previous sequence number that could have
been used since the other side will also create the flow anew on its
side. this last naxplanation could possibly be a resend, but we need
to make sure the other end gets it to not get stuck.

This also adds a dry mode to the local migration %mate task, that can
be used independently or together with the test %ahoy $plea.