mesa: fix migrating live naxplanations
this uses sequence numbers to account for naxplanation
pages for which we don't have any unacked messages—
this is the situation the flow is in when we are migrating
and need to peek for the naxplanation on the other side.

This could have the problem of resend naxplanation being
delivered twice? Another option would be to track this
naxplanation in state to certify that we were actually waiting
for them, instead of only allowing seq=(dec next.snd) to be
handled.