eyre: don't drop session expiry timer
Eyre's handling of session expiry timers was not correctly accounting
for the case where processing the timer crashed. It would print the
crash trace, but not take any other action. This means it would drop the
timer on the floor, resulting in stale sessions piling up in state.

Patching that bug itself is trivial, we simply emit a timer for a day
later. This leaves some stale sessions temporarily, but that buys the
ship time to get into a better state, where it hopefully can process the
timer correctly next time.

Cleaning up the consequences of the bug is less trivial. We need to set
up the timer if it has been dropped. This requires a scry into behn, and
trawling through all timers, so we don't want to do that unnecessarily.

Here, we add a flag into state that gets set to true when ships upgrade
into this new version of eyre. If the flag is true, we set it to false
and investigate whether we need to set up the timer again. If we do, we
set it, bringing us back into a normal state.